Create todo module (#10)
* fix logout issue by redirecting user to login page

* add todo create component

* finish todo components

* add task create and task list

* don't allow deleting a todo if it has existing tasks

* refresh todo after a delete

* add error interceptor

+++ server/src/todo/todo.service.ts
@@ -100,6 +100,13 @@ export class TodoService {
       );
     }
 
+    if (todo.tasks && todo.tasks.length > 0) {
+      throw new HttpException(
+        `Cannot delete this Todo list, it has existing tasks`,
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
     await this.todoRepo.delete({ id }); // delete todo list
 
     return toTodoDto(todo);
